refactor(executors): decompose grok-web into pure tool/markup leaves (1872→873)
Port of upstream diegosouzapw/OmniRoute#5994. Extracts pure tool-bridge,
native-tools, text-cleanup helpers and shared types into
open-sse/executors/grok-web/{tool-bridge,native-tools,text-cleanup,types}.ts;
the host imports them back. Preserves the fork-only generateTraceparent
header feature. Adds a split-boundary unit test.
